users (#5)
* Add user creation feature

Motivation for the change:
The motivation for this change is to introduce a user management feature to the application, allowing users to be created and managed within the system. This is a foundational step towards implementing user authentication and authorization.

How it differs from the previous implementation:
- Add a new `new_migration` task in `Taskfile.yaml` for creating database migrations.
- Modify the `run` task in `Taskfile.yaml` to use environment variables for database credentials.
- Add a new HTTP POST endpoint `/v1/users` in `endpoints.http` for creating users.
- Update `go.mod` to remove an indirect dependency comment.
- Update `postgres/db.go`, `postgres/models.go`, and `postgres/query.sql.go` to use `sqlc` version 1.27.0 and add new types and methods for handling nullable fields and user creation.
- Add a new `User` struct in `postgres/models.go` to represent user data.
- Add a new SQL query `CreateUser` in `sql/query/query.sql` for inserting new users into the database.
- Add new SQL migration files `000002_create_users_table.up.sql` and `000002_create_users_table.down.sql` to create and drop the `users` table.
- Update `sqlc.yaml` to version 2 and adjust configuration for code generation.
- Add a new route and handler `handleCreateUser` in `routes.go` for processing user creation requests, including password hashing using `bcrypt`.

* Add save recipe feature

Motivation for the change:
The motivation for this change is to allow users to save their favorite recipes to their profiles. This feature enhances user engagement by providing a personalized experience where users can easily access and manage their saved recipes.

How it differs from the previous implementation:
Previously, there was no functionality to save recipes to a user's profile. The new implementation introduces a new endpoint `/v1/save-recipe` that allows users to save recipes. It includes changes to the database schema to add a `saved_recipes` table, new SQL queries to handle saving and retrieving saved recipes, and updates to the server routes to handle the new endpoint. The `SavedRecipe` struct and related query methods are added to manage the saved recipes in the database.

* Add GET user endpoint with saved recipes

Motivation for the change:
The motivation for this change is to enhance the application's functionality by allowing users to retrieve their saved recipes. This feature improves user experience by providing easy access to their saved data.

How it differs from the previous implementation:
Previously, the application did not have an endpoint to get user details along with their saved recipes. The new implementation adds a GET endpoint `/v1/users/:id` that retrieves a user's information and their saved recipes. This involves adding a new handler function `handleGetUser` in `routes.go` and updating the `endpoints.http` file to include a sample request for this new endpoint.