Add digital sign feature and update documentation
Motivation for the change:
The motivation for this change is to introduce a new feature called "digital sign" to the application. This feature allows users to subscribe to a NATS topic and print messages to the console, enhancing the application's functionality and user interaction.

How it differs from the previous implementation:
- Add a new task `digitalsign` in the Taskfile to build and run the digital sign TUI application.
- Update the README.md to include instructions for creating a `docker.env` file and running the digital sign feature.
- Add a new image `digitalsign.png` to visually represent the digital sign feature in the documentation.
- Modify the .gitignore file to include `digitalsign` in the binary section.
- Reorganize and add new tasks in the Taskfile for better task management and clarity.