Add Kubernetes deployment instructions
Motivation for the change:
The motivation for this change is to provide users with the ability to deploy the application on a Kubernetes cluster. This addition enhances the flexibility and scalability of the application by allowing it to be deployed in a containerized environment.

How it differs from the previous implementation:
Previously, the README did not include any instructions or support for deploying the application on Kubernetes. The new implementation adds a section in the README with detailed instructions on how to deploy the application using Kubernetes deployment files. It also includes a new image file to visually support the instructions.