Refactor server and organize web package
Motivation for the change:
The motivation for this change is to improve the organization and modularity of the codebase by creating a dedicated web package. This helps in better managing the server-related code and makes it easier to maintain and extend in the future.

How it differs from the previous implementation:
- Move server-related files (http.go, middleware.go, pubsub_moq_test.go, routes.go, routes_test.go, server.go, testdata/create_recipe.json) into a new web package.
- Change the server initialization function from newServer to NewServer to follow Go naming conventions for exported functions.
- Update the server struct to export the Queries field, changing from queries to Queries.
- Introduce a new sql package to manage SQL schema files using Go's embed package for embedding SQL files.
- Update import paths and references throughout the code to reflect the new package structure.